Format contracts

+++ contracts/interfaces/IStaking.sol
@@ -2,11 +2,11 @@
 pragma solidity 0.8.12;
 
 interface IStaking {
-    function stake( uint _amount, address _recipient ) external returns ( bool );
+    function stake(uint256 _amount, address _recipient) external returns (bool);
 
-    function claim( address recipient ) external;
+    function claim(address recipient) external;
 
-    function unstake( uint _amount, bool _trigger ) external;
+    function unstake(uint256 _amount, bool _trigger) external;
 
-    function index() external view returns ( uint );
+    function index() external view returns (uint256);
 }
